Revert "Authentication & routing rework"
