+++ db/models.py
@@ -10,10 +10,3 @@ class Search(SQLModel, table=True):
     num_results: int
     timestamp: str
 
-class User(SQLModel, table=True):
-    id: Optional[int] = Field(default=None, primary_key=True)
-    username: str
-    password: str
-    email: str
-    user_type: str
-
