+++ requirements.txt
@@ -1,29 +1,19 @@
 anyio==3.6.2
-bcrypt==4.0.1
 beautifulsoup4==4.11.1
 bs4==0.0.1
 certifi==2022.9.24
-cffi==1.15.1
 charset-normalizer==2.1.1
 click==8.1.3
-cryptography==39.0.0
-ecdsa==0.18.0
 fastapi==0.85.1
+greenlet==1.1.3.post0
 h11==0.14.0
 idna==3.4
 importlib-metadata==5.0.0
 numpy==1.21.6
-passlib==1.7.4
 psycopg2-binary==2.9.4
-pyasn1==0.4.8
-pycparser==2.21
 pydantic==1.10.2
 python-dotenv==0.21.0
-python-jose==3.3.0
-python-multipart==0.0.5
 requests==2.28.1
-rsa==4.9
-six==1.16.0
 sniffio==1.3.0
 soupsieve==2.3.2.post1
 SQLAlchemy==1.4.41
